requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.20.0
loguru>=0.6.0
rich>=13.0.0
typing-extensions>=4.0.0
